requirements.txt
numpy==1.26.4
rasterio==1.3.11
PyYAML==6.0.2
pandas==2.2.2
matplotlib==3.9.0
elevation==1.1.3
mercantile==1.2.1
